Refeactor XP Fixes #396

+++ src/main/java/harmonised/pmmo/registry/LevelRegistry.java
@@ -10,17 +10,17 @@ import harmonised.pmmo.util.MsLoggy;
 import harmonised.pmmo.util.MsLoggy.LOG_CODE;
 
 public class LevelRegistry {
-	private final List<BiFunction<String, Integer, Integer>> providers = new ArrayList<>();
+	private final List<BiFunction<String, Long, Long>> providers = new ArrayList<>();
 	
-	public void registerLevelProvider(BiFunction<String, Integer, Integer> provider) {
+	public void registerLevelProvider(BiFunction<String, Long, Long> provider) {
 		Preconditions.checkNotNull(provider);
 		this.providers.add(provider);
 		MsLoggy.INFO.log(LOG_CODE.API, "Level Provider Registered");
 	}
 	
-	public int process(String skill, int nativeLevel) {
-		int outLevel = nativeLevel;
-		for (BiFunction<String, Integer, Integer> provider : providers) {
+	public long process(String skill, long nativeLevel) {
+		long outLevel = nativeLevel;
+		for (BiFunction<String, Long, Long> provider : providers) {
 			outLevel = provider.apply(skill, outLevel);
 		}
 		return outLevel;
